add python/ffmpeg convert anything to mp4 script (windows)
Requires Python 3. Just plop in the folder with the mixed files and run.
I will create an mp4 subfolder and put the converted files there. It's
up to you to clean the originals, but it will detect "already done"
files.

Speaking of, if there are incomplete conversions in the mp4 folder, they
will not get re-converted until the incomplete versions are removed.